% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Min-Max Scaling}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with values scaled between 0 and 1.
}
\description{
This function performs min-max scaling on a numeric vector, transforming the values to a range between 0 and 1.
}
\examples{
x <- c(1, 2, 3, 4, 5)
min_max_scale(x)
}
